Fix use-after-free segfault
During render, container nodes were created with a 'recursive free'
finalizer. As intermediate nodes were tossed away, their children would
then be freed multiple times.

The solution chosen here is to not use 'recursive free'. Instead,
there will always be one single foreign pointer per node. During render,
as the tree is built, the foreign pointers of the children will be
collected as references on the root. When the root is no longer live,
the children will naturally be destroyed as well.